avoid implicit type conversion

+++ src/tests/next/example/gallery/render/photon_mapping.cpp
@@ -324,7 +324,7 @@ int photon_mapping(Device &device) {
 
                 $if (cos_wi > 1e-4f) {
                     radiance = density_estimation_radius(coord, p, -ray->direction(), radius, material);
-                    radiance *= inv_pi / (radius * radius * photon_number);
+                    radiance *= inv_pi / (radius * radius * static_cast<float>(photon_number));
 
                     // $if(dot(radiance, radiance) > 10000) {
                     //     printer.info_with_location("p : ({}, {}, {}) o ({}, {}, {}) dir ({}, {}, {}) inst {} prim {}",
